refactor(data_table): delegate expand toggle to root controller
