+++ app/javascript/controllers/ruby_ui/data_table_controller.js
@@ -26,6 +26,17 @@ export default class extends Controller {
     this.updateState();
   }
 
+  toggleRowDetail(event) {
+    const button = event.currentTarget;
+    const id = button.getAttribute("aria-controls");
+    if (!id) return;
+    const target = document.getElementById(id);
+    if (!target) return;
+    const expanded = button.getAttribute("aria-expanded") === "true";
+    button.setAttribute("aria-expanded", String(!expanded));
+    target.classList.toggle("hidden", expanded);
+  }
+
   updateState() {
     const total = this.rowCheckboxTargets.length;
     const selected = this.rowCheckboxTargets.filter((cb) => cb.checked).length;
