+++ test/components/ruby_ui/data_table/data_table_expand_toggle_test.rb
@@ -1,14 +1,14 @@
 require "test_helper"
 
 class RubyUI::DataTableExpandToggleTest < ActiveSupport::TestCase
-  test "renders a button with aria attributes + controller" do
+  test "renders a button with aria attributes and delegated action" do
     out = render_component(RubyUI::DataTableExpandToggle.new(controls: "emp-1-detail"))
     assert_match(/<button[^>]*type="button"/, out)
     assert_match(/aria-expanded="false"/, out)
     assert_match(/aria-controls="emp-1-detail"/, out)
     assert_match(/aria-label="Toggle row details"/, out)
-    assert_match(/data-controller="[^"]*ruby-ui--data-table-row-expand/, out)
-    assert_match(/data-action="[^"]*click->ruby-ui--data-table-row-expand#toggle/, out)
+    assert_match(/data-action="[^"]*click->ruby-ui--data-table#toggleRowDetail/, out)
+    assert_no_match(/data-controller="ruby-ui--data-table-row-expand"/, out)
   end
 
   test "accepts custom label + initial expanded state" do
